share openContextMenu (#128065)
